Drop Google Tables and use Directus
