+++ map/data.csv
@@ -1057,7 +1057,7 @@ Flint,MI,US,FALSE,implemented,"add parking maximums,remove parking minimums",cit
 Flora,IL,US,FALSE,implemented,remove parking minimums,main street / special,commercial,,4803,38.6689364,-88.4856042,https://parkingreform.org/mandates-map/city_detail/Flora_IL.html,Samuel Deetz,"The area known as the Downtown Business Strip, described as businesses located on East North Avenue and West North Avenue between State Street and Sycamore Street, is exempted from minimum parking requirements."
 Florence,AL,US,FALSE,implemented,remove parking minimums,"city center / business district,main street / special",all uses,,40797,34.79981,-87.677251,https://parkingreform.org/mandates-map/city_detail/Florence_AL.html,Melissa Bailey,No parking is required in the Central Business District or the East Florence Historic District.
 Florence,CO,US,FALSE,implemented,"reduce parking minimums,remove parking minimums",city center / business district,"commercial,other",,3822,38.3902777,-105.1185988,https://parkingreform.org/mandates-map/city_detail/Florence_CO.html,Samuel Deetz,"Within the CB - Central Business district, and the CHB - Central Historic Business district, all uses except dwellings and lodging uses are exempt from parking mandates. Dwellings in these districts have reduced requirements."
-Florida,FL,US,FALSE,implemented,"reduce parking minimums,remove parking minimums",regional,all uses,,21538187,28.5421109,-81.3790304,https://parkingreform.org/mandates-map/city_detail/Florida_FL.html,,"In May 2024, Florida passed a bill to eliminate county and municipal parking requirements for transit oriented development, and reduce requirements near transit and public parking."
+Florida,FL,US,FALSE,implemented,"reduce parking minimums,remove parking minimums",regional,all uses,,21538187,27.9014,-81.5859,https://parkingreform.org/mandates-map/city_detail/Florida_FL.html,,"In May 2024, Florida passed a bill to eliminate county and municipal parking requirements for transit oriented development, and reduce requirements near transit and public parking."
 Florissant,MO,US,FALSE,implemented,add parking maximums,citywide,"commercial,industrial,medical,other",,52533,38.7919683,-90.3227803,https://parkingreform.org/mandates-map/city_detail/Florissant_MO.html,Samuel Deetz,Non-residential uses in all zoning districts may not contain more than one hundred ten percent (110%) of the minimum number of parking spaces required.
 Flushing,MI,US,FALSE,implemented,reduce parking minimums,city center / business district,all uses,,8411,43.0630834,-83.8510732,https://parkingreform.org/mandates-map/city_detail/Flushing_MI.html,Samuel Deetz,"Minimum parking requirements do not apply to uses in the Central Business District (CBD), at the discretion of the planning commission, if the commission determines that the parking needs of the use are met by any combination of on-street parking, joint use parking, parking on a nearby site and municipal off-street parking."
 Folcroft,PA,US,FALSE,implemented,reduce parking minimums,city center / business district,all uses,,6792,39.9005526,-75.279725,https://parkingreform.org/mandates-map/city_detail/Folcroft_PA.html,Samuel Deetz,"Uses in the GB General Business and RS Retail Service districts may use adjacent on-street and public parking spaces within 200 feet, if sufficient off-street parking is not available."
