+++ package.json
@@ -13,9 +13,6 @@
     "lint": "prettier --check . && eslint scripts/ tests/",
     "gen-csv": "tsx scripts/generateCsv.ts",
     "gen-html": "tsx scripts/generateHtmlPages.ts",
-    "sync-core-data": "tsx scripts/syncCoreData.ts",
-    "sync-extended-data": "tsx scripts/syncExtendedData.ts",
-    "sync-lat-lng": "tsx scripts/syncLatLng.ts",
     "sync-directus": "tsx scripts/syncDirectus.ts",
     "broken-links": "tsx scripts/brokenLinks.ts",
     "serve-dist": "cd dist; http-server",
